requirements.txt
paho-mqtt>=2.0.0
requests
psycopg2-binary
